Structure fix, minor bug fix, docker
